Add the GUI logic to chatbot.py and remove chatbotGUI.py. Modifies the logic to make it more like a typical chatbot interface
